CON-1199: Revert runtimeOnly change
`runtimeOnly` dependency on the enclave module does not work with CorDapps. The dependency needs to be declared as `compile` instead.

Unfortunately this is not picked up when running the tests, the Corda node has to be started manually for the issue to manifest. It seems the test classpath leaks in the enclave module which allows it to work.