CON-1241: Added PyTorch sample using WIP Gramine feature (#54)
Replaced the limited hello world Python sample with one that uses PyTorch to do image classification. This is possible with the very recent integration of Gramine into the Conclave SDK and the enabling of much better Python support.

Note, that this is still work in progress and is not yet ready for production use.